backend/requirements.txt
# FastAPI and server
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6

# LangChain and AI
langchain==0.1.4
langchain-google-genai==0.0.6
langchain-community==0.0.16

# Vector store and embeddings
faiss-cpu==1.7.4
sentence-transformers==2.2.2

# PDF processing
pypdf==4.0.1

# Environment and utilities
python-dotenv==1.0.0
